Extract Saver interface to allow it to be injected as dependency and ease testablility.
