+++ src/main/java/org/betonquest/betonquest/database/Saver.java
@@ -1,114 +1,38 @@
 package org.betonquest.betonquest.database;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import lombok.CustomLog;
-import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.database.Connector.UpdateType;
-import org.bukkit.Bukkit;
-import org.bukkit.event.Listener;
-
 import java.util.Arrays;
-import java.util.concurrent.ConcurrentLinkedQueue;
 
 /**
- * Saves the data to the database asynchronously
+ * The saver is used to save data via records into the database. Implementations should be thread save and must document
+ * if they are not.
  */
-@SuppressWarnings({"PMD.CommentRequired", "PMD.DoNotUseThreads"})
-@CustomLog
-@SuppressFBWarnings("IS2_INCONSISTENT_SYNC")
-public class Saver extends Thread implements Listener {
-
-    private final Connector con;
-    private final ConcurrentLinkedQueue<Record> queue;
-    private boolean running;
-
-    /**
-     * Creates new database saver thread
-     */
-    public Saver() {
-        super();
-        this.con = new Connector();
-        this.queue = new ConcurrentLinkedQueue<>();
-        this.running = true;
-        Bukkit.getPluginManager().registerEvents(this, BetonQuest.getInstance());
-    }
-
-    @Override
-    @SuppressFBWarnings("UW_UNCOND_WAIT")
-    public void run() {
-        boolean active = false;
-        while (true) {
-            while (queue.isEmpty()) {
-                if (!running) {
-                    return;
-                }
-                synchronized (this) {
-                    try {
-                        active = false;
-                        wait();
-                    } catch (final InterruptedException e) {
-                        LOG.error("There was a exception with SQL", e);
-                    }
-                }
-            }
-            if (!active) {
-                con.refresh();
-                active = true;
-            }
-            final Record rec = queue.poll();
-            con.updateSQL(rec.getType(), rec.getArgs());
-        }
-    }
-
+public interface Saver {
     /**
      * Adds new record to the queue, where it will be saved to the database.
      *
      * @param rec Record to save
      */
-    public void add(final Record rec) {
-        synchronized (this) {
-            queue.add(rec);
-            notifyAll();
-        }
-    }
+    void add(Record rec);
 
     /**
      * Ends this saver's job, letting it save all remaining data.
      */
-    public void end() {
-        synchronized (this) {
-            running = false;
-            notifyAll();
-        }
-    }
+    void end();
 
     /**
-     * Holds the data and the method of saving them to the database
+     * Holds the data and the method of saving them to the database.
      */
-    public static class Record {
-
-        private final UpdateType type;
-        private final String[] args;
-
+    record Record(Connector.UpdateType type, String... args) {
         /**
          * Creates new Record, which can be saved to the database using
-         * Saver.add()
+         * {@code Saver.add()}.
          *
          * @param type method used for saving the data
          * @param args list of Strings which will be saved to the database
          */
-        public Record(final UpdateType type, final String... args) {
+        public Record(final Connector.UpdateType type, final String... args) {
             this.type = type;
             this.args = args == null ? null : Arrays.copyOf(args, args.length);
         }
-
-        private UpdateType getType() {
-            return type;
-        }
-
-        private String[] getArgs() {
-            return args;
-        }
     }
-
 }
